+++ modules/generator/processor/spanmetrics/config.go
@@ -3,6 +3,7 @@ package spanmetrics
 import (
 	"flag"
 
+	"github.com/grafana/tempo/pkg/sharedconfig"
 	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
@@ -16,18 +17,24 @@ const (
 	dimSpanKind      = "span_kind"
 	dimStatusCode    = "status_code"
 	dimStatusMessage = "status_message"
+	dimJob           = "job"
+	dimInstance      = "instance"
 )
 
 type Config struct {
 	// Buckets for latency histogram in seconds.
 	HistogramBuckets []float64 `yaml:"histogram_buckets"`
 	// Intrinsic dimensions (labels) added to the metric, that are generated from fixed span
-	// data. The dimensions service, span_name, span_kind, and status_code are enabled by
+	// data. The dimensions service, span_name, span_kind, status_code, job and instance are enabled by
 	// default, whereas the dimension status_message must be enabled explicitly.
 	IntrinsicDimensions IntrinsicDimensions `yaml:"intrinsic_dimensions"`
 	// Additional dimensions (labels) to be added to the metric. The dimensions are generated
 	// from span attributes and are created along with the intrinsic dimensions.
 	Dimensions []string `yaml:"dimensions"`
+	// Dimension label mapping to allow the user to rename attributes in their metrics
+	DimensionMappings []sharedconfig.DimensionMappings `yaml:"dimension_mappings"`
+	// Enable target_info as a metrics
+	EnableTargetInfo bool `yaml:"enable_target_info"`
 
 	// If enabled attribute value will be used for metric calculation
 	SpanMultiplierKey string `yaml:"span_multiplier_key"`